+++ pages/signUpComplete/signUpComplete.css
@@ -1,48 +1,56 @@
 /*signUpComplete.css*/
 /*wrap*/
 #signup-complete-sction {
-    padding-top: 100px ;
-    padding-left: 100px ;
+  padding-top: 100px;
+  padding-left: 100px;
 }
 #signup-complete-sction #signup-complete-wrap {
-    /*border: 1px solid darkgrey;*/
-    width:400px;
+  /*border: 1px solid darkgrey;*/
+  width: 400px;
 }
 /*top*/
 #signup-complete-sction #signup-complete-wrap .title-wrap {
-    width:500px;
-    margin-bottom: 40px;
+  width: 500px;
+  margin-bottom: 40px;
 }
 
 #signup-complete-sction #signup-complete-wrap .title-wrap img.title {
-    max-width: 300px;
-    max-height: 100px;
-    margin-bottom: 10px;
-
+  max-width: 300px;
+  max-height: 100px;
+  margin-bottom: 10px;
 }
 
 #signup-complete-wrap .title-wrap .suv-title {
-    font-size: 40px;
+  font-size: 40px;
 }
 /*moveToMain*/
 #signup-complete-wrap .move-main-link {
-    color: #177edb;
-    font-size: 20px;
-    font-weight: bold;
-    margin-bottom: 25px;
+  margin-bottom: 25px;
 }
+#signup-complete-wrap .move-main-link a {
+  color: #177edb;
+  font-size: 1.5rem;
+  font-weight: bold;
+  text-decoration: none;
+}
+
 /*body box*/
 #signup-complete-sction #signup-complete-wrap .user-info-wrap {
-    border: 1px solid darkgrey;
-    padding: 15px;
-    margin-bottom: 10px;
+  border: 1px solid darkgrey;
+  padding: 15px;
+  margin-bottom: 10px;
 }
-#signup-complete-sction #signup-complete-wrap .user-info-wrap div {
-    color: grey;
-    margin-bottom: 10px;
+#signup-complete-sction #signup-complete-wrap .user-info-wrap .row {
+  margin-bottom: 10px;
+  display: flex;
+}
+#signup-complete-sction #signup-complete-wrap .user-info-wrap .key {
+  color: grey;
+  width: 50px;
+  margin-right: 10px;
 }
 /*small message*/
 #signup-complete-sction #signup-complete-wrap .small-messsage {
-    color: darkgrey;
-    font-size: 0.8rem;
+  color: darkgrey;
+  font-size: 0.8rem;
 }
