+++ pages/signUpComplete/signUpComplete.pug
@@ -5,7 +5,6 @@ html(leng='en')
         style
             include ./signUpComplete.css
             inlcude ../../styles/reset.css
-        script(type='text/javascript')
     body
         #signup-complete-sction
             #signup-complete-wrap
@@ -16,10 +15,20 @@ html(leng='en')
                     )
                     .suv-title 회원이 되셨습니다.
                 .move-main-link
-                    a 메인으로 ->
+                    a(href="/") 메인으로 &rarr;
                 .user-info-wrap
-                    .name 이름
-                    .id 아이디
-                    .email 이메일
-                    .phone 휴대폰
+                    .name.row
+                        .key 이름
+                        .value
+                    .id.row
+                        .key 아이디
+                        .value
+                    .email.row
+                        .key 이메일
+                        .value
+                    .phone.row
+                        .key 휴대폰
+                        .value
                 .small-messsage 위 내용은 내 정보관리에서 수정하실 수 있습니다.
+        script(type="module")
+            include ./signUpComplete.js
