--- types/userInfo.d.ts
@@ -1,6 +1,0 @@
-declare interface UserInfo {
-  id: string;
-  name: string;
-  email: string;
-  phoneNumber: string;
-}
